+++ formula-android/src/main/java/com/instacart/formula/android/events/ActivityResult.kt
@@ -1,4 +1,4 @@
-package com.instacart.formula.activity
+package com.instacart.formula.android.events
 
 import android.content.Intent